// src/lib/types.ts
import { LucideIcon } from "lucide-react";

export type Props = {};

export type Geo = {
  lat: string;
  lng: string;
};

export type Address = {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
};

export type Company = {
  name: string;
  catchPhrase: string;
  bs: string;
};

export type User = {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
};

export type NavProps = {
  isCollapsed: boolean;
  links: {
    title: string;
    label?: string;
    icon: LucideIcon;
    variant: "default" | "ghost";
    href: string;
  }[];
};

export type PageTitleProps = {
  title: string;
  classname?: string;
};

export type SalesProps = {
  name: string;
  email: string;
  salesAmount: string;
};